requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0
flask>=2.0.0
requests>=2.25.0

# Database and storage
firebase-admin>=5.0.0
pymongo>=4.0.0
boto3>=1.20.0

# Trading and exchange
python-binance>=1.0.16

# AI and ML
torch>=1.10.0
transformers>=4.12.0
nltk>=3.6.0
langchain-community>=0.0.10
openai>=1.0.0

# Text processing
beautifulsoup4>=4.10.0

# Time and date handling
pytz>=2021.3

# Web and API
Flask>=2.0.0
Werkzeug>=2.0.0

# Ollama for text summarization
# Note: Ollama itself is installed separately (see README) 